requirements.txt
streamlit
streamlit-webrtc
av
websockets<14.0
python-dotenv
nest_asyncio        # 追加
